ifab_chatbot_app/requirements.txt
sentence_transformers==2.7.0
streamlit
elasticsearch
openai
psycopg2-binary